fix bug

+++ BackdoorAttacks/utils.py
@@ -8,7 +8,7 @@ def store_indices(sorted_indices, harmful_which, dataset_name_clean, method_num,
     if poison_rate == None:
         path_sorted = os.path.join(directory, "sorted_{}_{}.json".format(method_num, dataset_name_clean))
     else:
-        path_sorted = os.path.join(directory, "sorted_{}_{}_{}%.json".format(method_num, dataset_name_clean,poison_rate*100))
+        path_sorted = os.path.join(directory, "sorted_{}_{}_{}%.json".format(method_num, dataset_name_clean, int(poison_rate*100)))
     path_hw = os.path.join(directory, "hw_{}_{}.json".format(method_num, dataset_name_clean))
 
     with open(path_sorted, "w") as json_file:
